Update Posns and Velocities to use doubles for smooth acceleration; Update main to demonstrate a player jumping off a ledge to show gravity; Tweak gravitational acceleration

+++ Portal-Plus-Plus/Portal-Plus-Plus/posn.cpp
@@ -1,10 +1,10 @@
 #include "posn.h"
 
-int Posn::getX() {
+double Posn::getX() {
 	return m_x;
 }
 
-int Posn::getY() {
+double Posn::getY() {
 	return m_y;
 }
